Aythentication

+++ frontend/src/components/BlogDetails.jsx
@@ -1,22 +1,31 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useBlogsContext } from "../hooks/useBlogsHook";
 
 //date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 export function BlogDetails({ blog }) {
   const {dispatch} = useBlogsContext();
+  const {user} = useAuthContext();
 
   const handleClick = async () => {
+    if(!user){
+      return;
+    }
+
     const response = await fetch("/api/blog/" + blog._id, {
       method: 'DELETE',
     });
     const json = await response.json();
-
     if(response.ok){
       dispatch({type: 'DELETE_BLOG', payload: json});
     }
+    
   }
+  
+
+
 
   return (
     <div className=" m-2 p-2 min-h-32 bg-white rounded">
@@ -27,7 +36,7 @@ export function BlogDetails({ blog }) {
       <p>
         {formatDistanceToNow(new Date(blog.createdAt), { addSuffix: true })}
       </p>
-      <p className="m-2"><strong>Created By:</strong>: {blog.Createdby?.email || 'Unknown'}</p>
+      <p className="m-2"><h3 className="font-semibold inline-block m-2">Created By : </h3>{blog.Createdby? blog.Createdby.email : 'Unknown'}</p>
       <button className="material-symbols-outlined ml-20" onClick={handleClick}>
         delete
       </button>
